backend/telegram-auth/requirements.txt
psycopg2-binary>=2.9.0
PyJWT>=2.8.0
bcrypt>=4.1.0
